pyproject.toml
[project]
name = "rocketrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie[model2vec]>=1.1.2",
    "fastapi>=0.116.1",
    "jinja2>=3.1.0",
    "kreuzberg>=3.11.2",
    "llama-cpp-python>=0.3.16",
    "numpy>=1.24.0",
    "pymilvus>=2.6.0",
    "rich>=13.0.0",
    "sentence-transformers>=5.1.0",
    "typer>=0.16.1",
    "uvicorn>=0.32.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rocketrag*"]
exclude = ["pdf*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]
viz = [
    "scikit-learn>=1.3.0",
]

[project.scripts]
rocketrag = "rocketrag.cli:main"
rag = "rocketrag.cli:main"
